реализован правки после мержа

+++ src/main/kotlin/ru/registrationbot/api/service/ClientService.kt
@@ -18,7 +18,7 @@ interface ClientService {
      * при удалении записи нужно оповестить клиента с помощью метода sendCancelNotification() класса RegistrationBot
      */
     //Возвращается chatId, если запись была найдена или null, чтобы можно было отправить сообщение клиенту
-    fun deleteRecording(idRecording: Long) : Long?
+    fun deleteRecording(idRecording: Long) : Int?
 
 
     /**
