+++ src/main/kotlin/ru/registrationbot/entities/ClientsEntity.kt
@@ -8,7 +8,7 @@ class ClientsEntity (
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    var id: Long? = null,
+    var id: Int? = null,
 
     @Column(name = "phone", nullable = false)
     var phone: String,
